feat(course): modularize RESTful API course and introduce core lesson engine
