+++ app/data/courses/restful-api-mastery-greybox/index.ts
@@ -2,6 +2,8 @@
 
 import type { Course } from '~/types/course'
 import { reiReltroner } from '~/data/authors'
+import { createCourse } from '~/data/core/course'
+import type { CourseInput } from '~/data/core/course'
 
 import module1 from './module-1-rest-anatomy'
 import module2 from './module-2-payload-meta'
@@ -10,34 +12,11 @@ import module4 from './module-4-crud-debug'
 import module5 from './module-5-automation'
 import module6 from './module-6-tool-agnostic'
 
-// 🔓 open lesson (sementara kosong, isi nanti di phase berikutnya)
 const OPEN_LESSON_SLUGS = new Set<string>([
-  'what-is-rest-and-greybox-thinking'
+  'what-is-rest-and-greybox-thinking',
 ])
 
-const modules = [
-  module1,
-  module2,
-  module3,
-  module4,
-  module5,
-  module6,
-].map(moduleItem => ({
-  ...moduleItem,
-  lessons: moduleItem.lessons.map(lesson => ({
-    ...lesson,
-    status: OPEN_LESSON_SLUGS.has(lesson.slug)
-      ? ('published' as const)
-      : ('draft' as const),
-  })),
-}))
-
-const lessonCount = modules.reduce(
-  (sum, moduleItem) => sum + moduleItem.lessons.length,
-  0,
-)
-
-const course: Course = {
+const baseCourse: CourseInput = {
   id: 'course-restful-api-mastery-greybox',
   slug: 'restful-api-mastery-greybox',
 
@@ -68,8 +47,14 @@ By the end of this course, you will not just "use APIs" — you will understand 
   difficulty: 'beginner',
   status: 'published',
 
-  lessonCount,
-  modules,
+  modules: [
+    module1,
+    module2,
+    module3,
+    module4,
+    module5,
+    module6,
+  ],
 
   tags: [
     'api',
@@ -86,7 +71,11 @@ By the end of this course, you will not just "use APIs" — you will understand 
   author: reiReltroner,
 
   createdAt: '2026-04-08',
-  updatedAt: '2026-04-08',
+  updatedAt: '2026-04-11',
 }
+
+const course = createCourse(baseCourse, {
+  openLessonSlugs: OPEN_LESSON_SLUGS,
+})
 
 export default course
