+++ app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/assignment-analyze-public-api.ts
@@ -1,0 +1,14 @@
+// app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/assignment-analyze-public-api.ts
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-1-5',
+  slug: 'assignment-analyze-public-api',
+  title: 'Assignment: Analyze Public API',
+  type: 'assignment',
+  order: 5,
+  durationMinutes: 20,
+  content: 'Analyze a public API and identify its structure and responses.',
+})
+
+export default lesson
