+++ app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/headers-basics.ts
@@ -1,0 +1,14 @@
+// app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/headers-basics.ts
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-1-4',
+  slug: 'headers-basics',
+  title: 'Headers Basics',
+  type: 'lesson',
+  order: 4,
+  durationMinutes: 8,
+  content: 'Explore request and response headers in API communication.',
+})
+
+export default lesson
