+++ app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/quiz-rest-fundamentals.ts
@@ -1,0 +1,14 @@
+// app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/quiz-rest-fundamentals.ts
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-1-6',
+  slug: 'quiz-rest-fundamentals',
+  title: 'Quiz: REST Fundamentals',
+  type: 'lesson',
+  order: 6,
+  durationMinutes: 10,
+  content: 'Test your understanding of REST fundamentals.',
+})
+
+export default lesson
