+++ app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/what-is-rest-and-greybox-thinking.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-1-rest-anatomy/what-is-rest-and-greybox-thinking.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-1-1',
+  slug: 'what-is-rest-and-greybox-thinking',
+  title: 'What is REST & Greybox Thinking',
+  type: 'lesson',
+  order: 1,
+  durationMinutes: 10,
+  content: 'Introduction to REST concepts and greybox thinking approach.',
+})
+
+export default lesson
