+++ app/data/courses/restful-api-mastery-greybox/module-3-security/assignment-secure-api-call.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-3-security/assignment-secure-api-call.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-3-4',
+  slug: 'assignment-secure-api-call',
+  title: 'Assignment: Secure API Call',
+  type: 'assignment',
+  order: 4,
+  durationMinutes: 20,
+  content: 'Implement a secure API call using authentication.',
+})
+
+export default lesson
