+++ app/data/courses/restful-api-mastery-greybox/module-3-security/authentication-vs-authorization.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-3-security/authentication-vs-authorization.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-3-1',
+  slug: 'authentication-vs-authorization',
+  title: 'Authentication vs Authorization',
+  type: 'lesson',
+  order: 1,
+  durationMinutes: 10,
+  content: 'Understand the difference between authentication and authorization.',
+})
+
+export default lesson
