+++ app/data/courses/restful-api-mastery-greybox/module-3-security/bearer-token-and-api-key.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-3-security/bearer-token-and-api-key.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-3-2',
+  slug: 'bearer-token-and-api-key',
+  title: 'Bearer Token & API Key',
+  type: 'lesson',
+  order: 2,
+  durationMinutes: 10,
+  content: 'Learn how API keys and bearer tokens work.',
+})
+
+export default lesson
