+++ app/data/courses/restful-api-mastery-greybox/module-3-security/common-security-pitfalls.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-3-security/common-security-pitfalls.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-3-3',
+  slug: 'common-security-pitfalls',
+  title: 'Common Security Pitfalls',
+  type: 'lesson',
+  order: 3,
+  durationMinutes: 12,
+  content: 'Explore common mistakes in API security.',
+})
+
+export default lesson
