+++ app/data/courses/restful-api-mastery-greybox/module-3-security/quiz-api-security.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-3-security/quiz-api-security.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-3-5',
+  slug: 'quiz-api-security',
+  title: 'Quiz: API Security',
+  type: 'lesson',
+  order: 5,
+  durationMinutes: 10,
+  content: 'Test your knowledge of API security concepts.',
+})
+
+export default lesson
