+++ app/data/courses/restful-api-mastery-greybox/module-4-crud-debug/assignment-debug-broken-api.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-4-crud-debug/assignment-debug-broken-api.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-4-4',
+  slug: 'assignment-debug-broken-api',
+  title: 'Assignment: Debug Broken API',
+  type: 'assignment',
+  order: 4,
+  durationMinutes: 20,
+  content: 'Fix issues in a broken API workflow.',
+})
+
+export default lesson
