+++ app/data/courses/restful-api-mastery-greybox/module-5-automation/quiz-automation-basics.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-5-automation/quiz-automation-basics.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-5-5',
+  slug: 'quiz-automation-basics',
+  title: 'Quiz: Automation Basics',
+  type: 'lesson',
+  order: 5,
+  durationMinutes: 10,
+  content: 'Test your automation knowledge.',
+})
+
+export default lesson
