+++ app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/cli-vs-gui-philosophy.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/cli-vs-gui-philosophy.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-6-1',
+  slug: 'cli-vs-gui-philosophy',
+  title: 'CLI vs GUI Philosophy',
+  type: 'lesson',
+  order: 1,
+  durationMinutes: 8,
+  content: 'Understand different approaches to working with APIs.',
+})
+
+export default lesson
