+++ app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/comparing-tools.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/comparing-tools.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-6-3',
+  slug: 'comparing-tools',
+  title: 'Comparing Tools (Postman vs CLI)',
+  type: 'lesson',
+  order: 3,
+  durationMinutes: 10,
+  content: 'Compare different tools for API testing.',
+})
+
+export default lesson
