+++ app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/final-reflection.ts
@@ -1,0 +1,15 @@
+// app/data/courses/restful-api-mastery-greybox/module-6-tool-agnostic/final-reflection.ts
+
+import { createLesson } from '~/data/core/lesson'
+
+const lesson = createLesson({
+  id: 'lesson-rest-6-5',
+  slug: 'final-reflection',
+  title: 'Final Reflection',
+  type: 'lesson',
+  order: 5,
+  durationMinutes: 10,
+  content: 'Reflect on your API learning journey.',
+})
+
+export default lesson
